backend/requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
google-genai
